requirements.txt
pytesseract==0.3.13
pillow==11.3.0
packaging==24.2
requests==2.32.5
typing-extensions==4.15.0
rich==14.3.3
pyttsx3==2.99
pynput==1.8.1
customtkinter==5.2.2
keyboard==0.13.5
python-dotenv==1.2.2
imagesize==2.0.0